feat: soc
